# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-grade e-commerce backend for OQart - India's premier organic products marketplace",
        "title": "OQart Backend API",
        "termsOfService": "https://oqart.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://oqart.com/support",
            "email": "support@oqart.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/vendors/documents/{id}/verify": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to verify or reject a vendor's uploaded document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Vendor Management"
                ],
                "summary": "Verify or reject vendor document (Admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Verification decision",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_internal_delivery_http_dto.AdminDocumentVerificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document verified successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/vendors/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin endpoint to approve, reject, or suspend a vendor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Vendor Management"
                ],
                "summary": "Approve or reject vendor (Admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vendor ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Approval decision",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_internal_delivery_http_dto.AdminVendorApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendor status updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Vendor not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/vendors": {
            "get": {
                "description": "Get a paginated list of vendors with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "List vendors",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, under_review, approved, rejected, suspended, inactive)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by verification status",
                        "name": "is_verified",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by featured status",
                        "name": "is_featured",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by business name or description",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_internal_delivery_http_dto.VendorListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vendors/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the authenticated vendor's complete profile with all details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor Management"
                ],
                "summary": "Get vendor dashboard",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_internal_delivery_http_dto.VendorDetailedResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Vendor not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the authenticated vendor's profile information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendor Management"
                ],
                "summary": "Update vendor profile",
                "parameters": [
                    {
                        "description": "Profile update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_internal_delivery_http_dto.UpdateVendorProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_internal_delivery_http_dto.VendorResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input or validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Vendor not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_oqart_backend_pkg_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_oqart_backend_internal_delivery_http_dto.AdminDocumentVerificationRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "rejection_reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "approved",
                        "rejected"
                    ]
                }
            }
        },
        "github_com_oqart_backend_internal_delivery_http_dto.AdminVendorApprovalRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "rejection_reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "approved",
                        "rejected",
                        "suspended"
                    ]
                }
            }
        },
...
        "github_com_oqart_backend_internal_delivery_http_dto.UpdateVendorProfileRequest": {
            "type": "object",
            "properties": {
                "banner_url": {
                    "type": "string"
                },
                "business_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "logo_url": {
                    "type": "string"
                }
            }
        },
...
        "github_com_oqart_backend_internal_delivery_http_dto.VendorListResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "vendors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_oqart_backend_internal_delivery_http_dto.VendorResponse"
                    }
                }
            }
        },
...
        "github_com_oqart_backend_pkg_errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}